+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- Updated Return Name Table to use left-hand arrow button for expand/collapse of chart
+
 ### Fixed
 
 - Fixed funding source option for primary name requests.